Seperating dialog test app
